Trimtokyo/requirements.txt
Flask
Flask-Mail
python-dotenv
WeasyPrint
gunicorn
firebase-admin
requests
